Restructure integration test
Make it easier to define multiple tests and run them concurrently.
