+++ integration/nginx.conf
@@ -15,31 +15,20 @@ http {
     error_log stderr;
 
     init_worker_by_lua_block {
-        prometheus = require("prometheus").init("prometheus_metrics",
-	  {sync_interval=0.4})
+        prometheus = require("prometheus").init("prometheus_metrics", {sync_interval=0.4})
         metric_requests = prometheus:counter("requests_total",
-          "Number of HTTP requests", {"host", "status"})
+          "Number of HTTP requests", {"host", "path", "status"})
         metric_latency = prometheus:histogram("request_duration_seconds",
-          "HTTP request latency", {"host"},
+          "HTTP request latency", {"path"},
           {0.08, 0.089991, 0.1, 0.2, 0.75, 1, 1.5, 3.123232001, 5, 15, 120, 350.5, 1500, 75000, 1500000})
         metric_connections = prometheus:gauge("connections",
           "Number of HTTP connections", {"state"})
     }
-    log_by_lua_block {
-        metric_requests:inc(1, {ngx.var.server_name, ngx.var.status})
-        metric_latency:observe(tonumber(ngx.var.request_time),
-                               {ngx.var.server_name})
-    }
 
     server {
-        listen 18001;
-        server_name fast;
-        root /nginx-lua-prometheus/integration/www;
-        index index.txt;
+        listen 18000;
+        server_name metrics;
 
-        location /error {
-            return 500;
-        }
         location /metrics {
             content_by_lua_block {
                 metric_connections:set(ngx.var.connections_reading, {"reading"})
@@ -49,15 +38,31 @@ http {
             }
         }
     }
+
     server {
-        listen 18002;
-        server_name slow;
-        location / {
+        listen 18001;
+        server_name basic_test;
+
+        log_by_lua_block {
+            metric_requests:inc(1, {ngx.var.server_name, ngx.var.request_uri, ngx.var.status})
+            metric_latency:observe(tonumber(ngx.var.request_time),
+                                {ngx.var.request_uri})
+        }
+
+        location /fast {
+            content_by_lua_block {
+                ngx.say("ok")
+            }
+        }
+        location /slow {
             content_by_lua_block {
                 -- sleep for 10-20ms.
                 ngx.sleep(0.01 + (0.01 * math.random()))
                 ngx.say("ok")
             }
+        }
+        location /error {
+            return 500;
         }
     }
 }
