+++ integration/test.go
@@ -2,111 +2,35 @@
 package main
 
 import (
+	"context"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
-	"math"
-	"math/rand"
 	"net/http"
 	"sync"
 	"time"
 
-	"github.com/golang/protobuf/proto"
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/kr/pretty"
 	dto "github.com/prometheus/client_model/go"
 	"github.com/prometheus/common/expfmt"
 )
 
 var (
 	testDuration = flag.Duration("duration", 10*time.Second, "duration of the test")
-	concurrency  = flag.Int("concurrency", 9, "number of concurrent http clients")
 )
 
-type requestType int
+const metricsURL = "http://localhost:18000/metrics"
 
-const (
-	reqFast requestType = iota
-	reqSlow
-	reqError
-)
-
-// There are three separate nginx endpoints:
-// - 'fast' simply returns "ok" with a 200 response code;
-// - 'slow' waits for 10ms and returns "ok" with a 200 response code;
-// - 'error' returns a 500.
-var urls = map[requestType]string{
-	reqFast:  "http://localhost:18001/",
-	reqSlow:  "http://localhost:18002/",
-	reqError: "http://localhost:18001/error",
+type testRunner struct {
+	client *http.Client
+	tests  []testFunc
+	checks []checkFunc
 }
 
-// Verify bucket boundaries. This should match the buckets defined in nginx.conf.
-var buckets = []float64{0.08, 0.089991, 0.1, 0.2, 0.75, 1, 1.5, 3.123232001, 5, 15, 120, 350.5, 1500, 75000, 1500000, math.Inf(1)}
+type testFunc func() error
+type checkFunc func(*testData) error
 
-// getHistogramSum returns the 'sum' value for a given histogram metric.
-func getHistogramSum(mfs map[string]*dto.MetricFamily, metric string, labels [][]string) float64 {
-	var lps []*dto.LabelPair
-	for _, lp := range labels {
-		lps = append(lps, &dto.LabelPair{Name: proto.String(lp[0]), Value: proto.String(lp[1])})
-	}
-
-	for _, mf := range mfs {
-		if *mf.Name == metric {
-			for _, m := range mf.Metric {
-				if cmp.Equal(m.Label, lps) {
-					return *m.Histogram.SampleSum
-				}
-			}
-		}
-	}
-	log.Fatalf("Metric %s with labels %v not found in %v", metric, lps, mfs)
-	return 0
-}
-
-// hasMetricFamily verifies that a given MetricFamily exists in a passed list of
-// metric families.
-func hasMetricFamily(mfs map[string]*dto.MetricFamily, want *dto.MetricFamily) error {
-	sortFn := func(x, y interface{}) bool { return pretty.Sprint(x) < pretty.Sprint(y) }
-	for _, mf := range mfs {
-		if *mf.Name == *want.Name {
-			if diff := cmp.Diff(want, mf, cmpopts.SortSlices(sortFn)); diff != "" {
-				return fmt.Errorf("Unexpected metric family %v (-want +got):\n%s", mf.Name, diff)
-			}
-			return nil
-		}
-	}
-	return fmt.Errorf("Metric family %v not found in %v", want, mfs)
-}
-
-// checkBucketBoundaries verifies bucket boundary values.
-func checkBucketBoundaries(mfs map[string]*dto.MetricFamily, metric string) error {
-	matched := false
-	for _, mf := range mfs {
-		if *mf.Name != metric {
-			continue
-		}
-		matched = true
-		for _, m := range mf.Metric {
-			if len(m.Histogram.Bucket) != len(buckets) {
-				return fmt.Errorf("expected %d buckets but got %d: %v", len(buckets), len(m.Histogram.Bucket), m.Histogram.Bucket)
-			}
-			for idx, b := range m.Histogram.Bucket {
-				tolerance := 0.00001
-				if diff := math.Abs(*b.UpperBound - buckets[idx]); diff > tolerance {
-					return fmt.Errorf("unexpected value for bucket #%d; want %f got %f", idx, buckets[idx], *b.UpperBound)
-				}
-			}
-		}
-	}
-
-	if !matched {
-		return fmt.Errorf("could not find metric %s", metric)
-	}
-
-	return nil
+type testData struct {
+	metrics map[string]*dto.MetricFamily
 }
 
 func main() {
@@ -118,136 +42,76 @@ func main() {
 		Transport: &http.Transport{IdleConnTimeout: 400 * time.Millisecond},
 	}
 
-	log.Printf("Starting the test with %d concurrent clients", *concurrency)
+	// Wait for nginx to start.
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+	if err := waitFor(ctx, client, metricsURL); err != nil {
+		log.Fatal(err)
+	}
+
+	// Register tests.
+	t := &testRunner{client: client}
+	registerBasic(t)
+
+	// Run tests.
 	var wg sync.WaitGroup
-	results := make(chan map[requestType]int64, *concurrency)
-	for i := 1; i <= *concurrency; i++ {
+	for _, tt := range t.tests {
 		wg.Add(1)
-		go func() {
-			result := make(map[requestType]int64)
-			for start := time.Now(); time.Since(start) < *testDuration; {
-				t := reqFast
-				r := rand.Intn(100)
-				if r < 10 {
-					// 10% are slow requests
-					t = reqSlow
-				} else if r < 15 {
-					// 5% are errors
-					t = reqError
-				}
-				resp, err := client.Get(urls[t])
-				if err != nil {
-					log.Fatalf("Could not fetch URL %s: %v", urls[t], err)
-				}
-				body, err := ioutil.ReadAll(resp.Body)
-				if err != nil {
-					log.Fatalf("Could not read HTTP response for %s: %v", urls[t], err)
-				}
-				resp.Body.Close()
-				if t != reqError && string(body) != "ok\n" {
-					log.Fatalf("Unexpected response %q from %s; expected 'ok'", string(body), urls[t])
-				}
-				result[t]++
+		go func(tt testFunc) {
+			if err := tt(); err != nil {
+				log.Fatal(err)
 			}
-			results <- result
 			wg.Done()
-		}()
+		}(tt)
 	}
 	wg.Wait()
-	close(results)
-
-	var fast, slow, errors int64
-	for r := range results {
-		fast += r[reqFast]
-		slow += r[reqSlow]
-		errors += r[reqError]
-	}
-	total := fast + slow + errors
-	log.Printf("Sent %d requests (%d fast, %d slow, %d errors)", total, fast, slow, errors)
 
 	// Sleep for 500ms before collecting metrics. This is to ensure that all HTTP connections
 	// to nginx get closed, and to allow for some eventual consistency in nginx-lua-prometheus.
 	time.Sleep(500 * time.Millisecond)
 
-	resp, err := client.Get("http://localhost:18001/metrics")
+	// Collect metrics.
+	resp, err := client.Get(metricsURL)
 	if err != nil {
 		log.Fatalf("Could not collect metrics: %v", err)
 	}
 	defer resp.Body.Close()
 
+	// Parse metrics.
 	var parser expfmt.TextParser
-	mfs, err := parser.TextToMetricFamilies(resp.Body)
+	res := &testData{}
+	res.metrics, err = parser.TextToMetricFamilies(resp.Body)
 	if err != nil {
 		log.Fatalf("Could not parse metrics: %v", err)
 	}
 
-	// We expect all fast requests to take less than 1 second.
-	if v := getHistogramSum(mfs, "request_duration_seconds", [][]string{{"host", "fast"}}); v > 1 {
-		log.Fatalf("Total time to process all fast request is %f; expected <= 1", v)
-	}
-
-	minSlowSeconds := float64(slow) * 0.01 // at least 10ms per request
-	if v := getHistogramSum(mfs, "request_duration_seconds", [][]string{{"host", "slow"}}); v <= minSlowSeconds {
-		log.Fatalf("Total time to process all fast request is %f; expected > %f", v, minSlowSeconds)
-	}
-
-	expected := []*dto.MetricFamily{
-		{
-			// There should be no errors reported by the library.
-			Name:   proto.String("nginx_metric_errors_total"),
-			Help:   proto.String("Number of nginx-lua-prometheus errors"),
-			Type:   dto.MetricType_COUNTER.Enum(),
-			Metric: []*dto.Metric{{Counter: &dto.Counter{Value: proto.Float64(0)}}},
-		},
-		{
-			Name: proto.String("requests_total"),
-			Help: proto.String("Number of HTTP requests"),
-			Type: dto.MetricType_COUNTER.Enum(),
-			Metric: []*dto.Metric{
-				{Label: []*dto.LabelPair{
-					{Name: proto.String("host"), Value: proto.String("fast")},
-					{Name: proto.String("status"), Value: proto.String("200")},
-				}, Counter: &dto.Counter{Value: proto.Float64(float64(fast))}},
-				{Label: []*dto.LabelPair{
-					{Name: proto.String("host"), Value: proto.String("slow")},
-					{Name: proto.String("status"), Value: proto.String("200")},
-				}, Counter: &dto.Counter{Value: proto.Float64(float64(slow))}},
-				{Label: []*dto.LabelPair{
-					{Name: proto.String("host"), Value: proto.String("fast")},
-					{Name: proto.String("status"), Value: proto.String("500")},
-				}, Counter: &dto.Counter{Value: proto.Float64(float64(errors))}},
-			},
-		},
-		{
-			// After all tests are complete there should only be a single HTTP
-			// connection, which should be in 'writing' state (the connection
-			// used by nginx to return metric values).
-			Name: proto.String("connections"),
-			Help: proto.String("Number of HTTP connections"),
-			Type: dto.MetricType_GAUGE.Enum(),
-			Metric: []*dto.Metric{
-				{Label: []*dto.LabelPair{
-					{Name: proto.String("state"), Value: proto.String("reading")},
-				}, Gauge: &dto.Gauge{Value: proto.Float64(float64(0))}},
-				{Label: []*dto.LabelPair{
-					{Name: proto.String("state"), Value: proto.String("waiting")},
-				}, Gauge: &dto.Gauge{Value: proto.Float64(float64(0))}},
-				{Label: []*dto.LabelPair{
-					{Name: proto.String("state"), Value: proto.String("writing")},
-				}, Gauge: &dto.Gauge{Value: proto.Float64(float64(1))}},
-			},
-		},
-	}
-
-	for _, mf := range expected {
-		if err := hasMetricFamily(mfs, mf); err != nil {
+	// Run test checks.
+	for _, ch := range t.checks {
+		if err := ch(res); err != nil {
 			log.Fatal(err)
 		}
 	}
 
-	if err := checkBucketBoundaries(mfs, "request_duration_seconds"); err != nil {
-		log.Fatal(err)
-	}
-
 	log.Print("All ok")
+}
+
+func waitFor(ctx context.Context, c *http.Client, url string) error {
+	log.Printf("Waiting for %s...", url)
+	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
+	if err != nil {
+		return fmt.Errorf("creating request for %s: %v", url, err)
+	}
+	for {
+		resp, err := c.Do(req)
+		if err == nil {
+			resp.Body.Close()
+			if resp.StatusCode == 200 {
+				return nil
+			}
+		}
+		if err := ctx.Err(); err != nil {
+			return err
+		}
+		time.Sleep(100 * time.Millisecond)
+	}
 }
